% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHL26.model.coef.df}
\alias{CHL26.model.coef.df}
\title{CHL26 Model Coefficients}
\format{A data.frame with 2 variables: \code{geneName}, \code{coef}}
\usage{
CHL26.model.coef.df
}
\description{
Predictor feature coefficients as published in paper.
}
\keyword{datasets}

