% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHL26.test.exprs.mat}
\alias{CHL26.test.exprs.mat}
\title{CHL26 Test Expression Matrix}
\format{A 26 by 10 matrix}
\usage{
CHL26.test.exprs.mat
}
\description{
A synthetically generated expression matrix to be used to test the package
}
\keyword{datasets}

