% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_RCC_to_mat.R
\name{convert_RCC_to_mat}
\alias{convert_RCC_to_mat}
\title{Convert Nanostring RCC Files to Matrix Form}
\usage{
convert_RCC_to_mat(rcc.dir)
}
\arguments{
\item{rcc.dir}{The path to the directory holding the RCC files}
}
\value{
Matrix containing the raw nanostring count data of each RCC file
}
\description{
\code{convert_RCC_to_mat} converts the RCC files in a directory into a
matrix.
}
\author{
Stacy Hung
}

