% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_CHL26_scores.R
\name{get_CHL26_scores}
\alias{get_CHL26_scores}
\title{Generate the CHL 26 Predictor Scores}
\usage{
get_CHL26_scores(mat.norm)
}
\arguments{
\item{mat.norm}{Normalized expression matrix generated by
\code{normalizer_mat}}
}
\value{
A two column data.frame with sampleID and scores
}
\description{
\code{get_CHL26_score} returns the CHL26 score
}
\examples{
normalizer <- get_normalizer(mat)
mat.norm <- normalize_mat(mat, normalizer)
get_CHL26_scores(mat.norm)
}

