% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/melt_mat.R
\name{melt_mat}
\alias{melt_mat}
\title{Melts the Expression Matrix}
\usage{
melt_mat(in.mat)
}
\arguments{
\item{in.mat}{Gene expression}
}
\value{
data.frame of the melted gene expression matrix
}
\description{
Uses the reshape2::melt() function to melt the gene expression matrix
to return a sample-gene expression data.frame
}

