% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize_mat.R
\name{normalize_mat}
\alias{normalize_mat}
\title{Normalize the Expression Matrix}
\usage{
normalize_mat(mat, normalizer)
}
\arguments{
\item{mat}{The expression matrix}

\item{normalizer}{Normalizer values generated by \code{get_normalizer}}
}
\value{
A normalized expression matrix
}
\description{
This function normalizes the expression matrix.
}
\examples{
normalizer <- get_normalizer(mat)
mat.norm <- normalizer_mat(mat, normalizer)
}

